refactor: web & mobile now both use MSW
Mock Service Worker (MSW) is an API mocking library. With MSW we can
intercept outgoing requests, observe them, and respond to them using
mocked responses. This simplifies our tests a lot.

Up until now we were mocking the method that would do a network request.
This always required knowledge of the internal workings of our APIs.

MSW runs in the background and outputs a warning in the terminal when we
are trying to access a network resource for which we haven’t mocked a
response. This way we only need to provide dummy data for our endpoints
instead of fiddling with internal logic.

Updating to MSW outlined problems in JSDOM and I had to update to
jest-fixed-jsdom. We were manually fixing some of those problems like
missing global fetch, response and request objects. Now our jsdom
environment has the global objects that are available both in the
browser and in node, so we don’t introduce pollyfills and maintain them.